# perf: refactor get_experiments.sql for a better query plan